+++ src/types.ts
@@ -126,6 +126,7 @@ export interface I18nextToolkitConfig {
     /**
      * The format of the output translation files.
      * 'json': Standard JSON file (default)
+     * 'json5': JSON5 file
      * 'js': JavaScript file with ES Module syntax (export default)
      * 'js-esm': JavaScript file with ES Module syntax (export default)
      * 'js-cjs': JavaScript file with CommonJS syntax (module.exports)